only enable feature if enabled in configuration file

+++ vppdhc/main.py
@@ -19,22 +19,25 @@ def version_callback(value: bool):
 
 async def setup_tasks(conf, vpp):
     # DHCPv4 server
-    c = conf['dhc4server']
-    if os.path.exists(c['socket']):
-        os.remove(c['socket'])
+    tasks = []
+    if 'dhc4server' in conf
+        c = conf['dhc4server']
+        if os.path.exists(c['socket']):
+            os.remove(c['socket'])
 
-    dhcp_server = DHCPServer(c['socket'], conf['vpp']['socket'], vpp,
-                             c['renewal-time'], c['lease-time'], c['name-server'])
-
-    # DHCPv6 PD client
-    c = conf['dhc6pdclient']
-    if os.path.exists(c['socket']):
-        os.remove(c['socket'])
-    pd_client = DHCPv6PDClient(c['socket'], conf['vpp']['socket'],
-                               vpp,
-                               c['interface'], c['internal-prefix'])
-
-    await asyncio.gather(dhcp_server(), pd_client())
+        dhcp_server = DHCPServer(c['socket'], conf['vpp']['socket'], vpp,
+                                c['renewal-time'], c['lease-time'], c['name-server'])
+        tasks.append(dhcp_server())
+    if 'dhc6server' in conf:
+        # DHCPv6 PD client
+        c = conf['dhc6pdclient']
+        if os.path.exists(c['socket']):
+            os.remove(c['socket'])
+        pd_client = DHCPv6PDClient(c['socket'], conf['vpp']['socket'],
+                                vpp,
+                                c['interface'], c['internal-prefix'])
+        tasks.append(pd_client())
+    await asyncio.gather(tasks)
 
 @app.command()
 def main(config: typer.FileText,
